+++ src/emu/machines/zxSpectrumP3e/ZxSpectrumP3eMachine.ts
@@ -3,6 +3,7 @@ import type { ISpectrumPsgDevice } from "@emu/machines/zxSpectrum/ISpectrumPsgDe
 import type { CodeInjectionFlow } from "@emu/abstractions/CodeInjectionFlow";
 import type { MachineModel } from "@common/machines/info-types";
 import type { IFloppyControllerDevice } from "@emu/abstractions/IFloppyControllerDevice";
+import type { AudioSample } from "@emu/abstractions/IAudioDevice";
 
 import { TapeMode } from "@emu/abstractions/TapeMode";
 import { SpectrumBeeperDevice } from "../BeeperDevice";
@@ -362,13 +363,16 @@ export class ZxSpectrumP3EMachine extends ZxSpectrumBase {
    * Gets the audio samples rendered in the current frame
    * @returns Array with the audio samples
    */
-  getAudioSamples(): number[] {
+  getAudioSamples(): AudioSample[] {
     const beeperSamples = this.beeperDevice.getAudioSamples();
     const psgSamples = this.psgDevice.getAudioSamples();
     const samplesCount = Math.min(beeperSamples.length, psgSamples.length);
-    var sumSamples: number[] = [];
+    const sumSamples: AudioSample[] = [];
     for (let i = 0; i < samplesCount; i++) {
-      sumSamples[i] = beeperSamples[i] + psgSamples[i];
+      sumSamples[i] = {
+        left: beeperSamples[i].left + psgSamples[i].left,
+        right: beeperSamples[i].right + psgSamples[i].right
+      };
     }
     return sumSamples;
   }